fix(wcbet): resolve pyright type errors across the betting cog
Type self.bot as Hablemos (via TYPE_CHECKING + from __future__ import
annotations) so .db/.settings resolve, add WcBet/WcWallet TypedDicts for
the bet/wallet row shapes, and narrow Optional guild/author/channel
accesses at the call sites with explicit guards.

Also fix a genuine signature bug: place_wc_bet typed odds as float, but
the column is NUMERIC and every caller passes Decimal — correct it to
Decimal and update the one test that passed a float literal.

Clears all 70 pyright errors in cogs/wcbet_cog (507 -> 437 project-wide).
No behavior change; wcbet tests and smoke tests pass.

+++ db/bets.py
@@ -193,7 +193,7 @@ class WCBetsMixin(DatabaseMixin):
         match_id: int,
         outcome: str,
         stake: int,
-        odds: float,
+        odds: Decimal,
     ) -> int:
         """Place (or replace) the user's bet on a match in one transaction.
